working on edit function
