+++ src/Components/HomeLogin/HomeLogin.scss
@@ -4,6 +4,10 @@
 // Login form for mobile
 .form {
   padding: 0 1rem;
+  width: 100%;
+  height: 100%;
+  margin: auto;
+  text-align: center;
   //  display: flex;
   //   justify-content: center;
 
@@ -66,6 +70,8 @@
     display: flex;
     align-items: center;
     justify-content: center;
+    text-align: center;
+    margin: auto;
     padding: 1rem;
     height: 2rem;
     background-color: white;
